feat: move util module into sql_app
